修正用qs启动时没有拉起nghttpx
